Added loading screen that checks if user is logged in. It opens log on screen if not, and map screen if yes. Added log on screen where user enters phone number and receives verification code. Added dummy method for verification if user is logged in. Moved main logic from main.dart to map_screen.dart. Changed all colors so that now they are controlled only in two variables in main.dart and color scheme of entire app can be easily changed.
