+++ lib/src/screens/login_screen.dart
@@ -1,0 +1,184 @@
+import 'package:flutter/material.dart';
+
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({super.key});
+
+  @override
+  _LoginScreenState createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
+  final TextEditingController _phoneNumberController = TextEditingController();
+  final TextEditingController _verificationCodeController = TextEditingController();
+
+  String _selectedLanguage = 'English';
+  bool _termsAccepted = false;
+
+  void _verifyPhoneNumber(BuildContext context) {
+    // Perform phone number verification logic here
+    // You can use a service like Firebase Authentication for this
+
+    // Simulating successful verification for demonstration
+    String verificationCode = "111111";
+
+    if (_verificationCodeController.text == verificationCode) {
+      Navigator.pushNamed(
+        context,
+        '/map',
+      );
+    } else {
+      showDialog(
+        context: context,
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: const Text('Verification Failed'),
+            content: const Text('Please enter a valid verification code.'),
+            actions: <Widget>[
+              TextButton(
+                child: const Text('OK'),
+                onPressed: () {
+                  Navigator.of(context).pop();
+                },
+              ),
+            ],
+          );
+        },
+      );
+    }
+  }
+
+  void _showVerificationDialog(BuildContext context) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Verification Code'),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              const Text('Please enter the 6-digit verification code:'),
+              const SizedBox(height: 16.0),
+              TextFormField(
+                controller: _verificationCodeController,
+                keyboardType: TextInputType.number,
+                maxLength: 6,
+                decoration: const InputDecoration(
+                  labelText: 'Verification Code',
+                ),
+              ),
+            ],
+          ),
+          actions: <Widget>[
+            TextButton(
+              child: const Text('OK'),
+              onPressed: () {
+                if (_formKey.currentState!.validate()) {
+                  _verifyPhoneNumber(context);
+                }
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Login Page'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Form(
+          key: _formKey,
+          child: Column(
+            children: <Widget>[
+              TextFormField(
+                controller: _phoneNumberController,
+                keyboardType: TextInputType.phone,
+                decoration: const InputDecoration(
+                  labelText: 'Phone Number',
+                ),
+                validator: (String? value) {
+                  if (value == null || value.isEmpty) {
+                    return 'Please enter phone number';
+                  }
+                  return null;
+                },
+              ),
+              const SizedBox(height: 16.0),
+              DropdownButtonFormField<String>(
+                value: _selectedLanguage,
+                decoration: const InputDecoration(
+                  labelText: 'Language',
+                ),
+                items: <String>['English', 'Polish'].map<DropdownMenuItem<String>>((String value) {
+                  return DropdownMenuItem<String>(
+                    value: value,
+                    child: Text(value),
+                  );
+                }).toList(),
+                onChanged: (String? newValue) {
+                  setState(() {
+                    _selectedLanguage = newValue!;
+                  });
+                },
+                validator: (String? value) {
+                  if (value == null) {
+                    return 'Please select a language';
+                  }
+                  return null;
+                },
+              ),
+              const SizedBox(height: 16.0),
+              Row(
+                children: <Widget>[
+                  Checkbox(
+                    value: _termsAccepted,
+                    onChanged: (bool? newValue) {
+                      setState(() {
+                        _termsAccepted = newValue!;
+                      });
+                    },
+                  ),
+                  const Text('I accept the terms and conditions'),
+                ],
+              ),
+              const SizedBox(height: 16.0),
+              ElevatedButton(
+                onPressed: () {
+                  if (!_termsAccepted) {
+                    showDialog(
+                      context: context,
+                      builder: (BuildContext context) {
+                        return AlertDialog(
+                          title: const Text('Terms and Conditions'),
+                          content: const Text(
+                              'Please accept the terms and conditions.'),
+                          actions: <Widget>[
+                            TextButton(
+                              child: const Text('OK'),
+                              onPressed: () {
+                                Navigator.of(context).pop();
+                              },
+                            ),
+                          ],
+                        );
+                      },
+                    );
+                  } else if (_formKey.currentState!.validate()) {
+                    _showVerificationDialog(context);
+                  }
+                },
+                child: const Text('Verify Phone Number'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
