+++ lib/src/screens/user_registration.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+/*import 'package:flutter/material.dart';
 
 class RegistrationForm extends StatefulWidget {
   @override
@@ -112,3 +112,4 @@ class _PhoneVerificationScreenState extends State<PhoneVerificationScreen> {
     // Handle the response as per your API's requirements
   }
 }
+*/
