+++ lib/src/widgets/attendees_list.dart
@@ -14,13 +14,10 @@ class AttendeesList extends StatefulWidget {
 
 class _AttendeesListState extends State<AttendeesList> {
   late List<dynamic> attendeesList = widget.attendeesList;
-  Color mainColor = Colors.orange.shade700;
-  Color secondaryColor = Colors.orange.shade600;
 
   @override
   Widget build(BuildContext context) {
     return Expanded(
-      //crossAxisAlignment: CrossAxisAlignment.start,
       child:
         Wrap(
           spacing: 2.0,
@@ -48,8 +45,6 @@ class _AttendeesListState extends State<AttendeesList> {
         }(),
         style: const TextStyle(color: Colors.white, fontSize: 10),
       ),
-      backgroundColor: mainColor,
-      deleteIconColor: Colors.white,
       padding: const EdgeInsets.all(0.0),
       labelPadding: const EdgeInsets.only(top: -4.0, bottom: -4.0, left: 6.0),
       onDeleted: () {
