--- lib/src/widgets/login_form.dart
@@ -1,77 +1,0 @@
-/*import 'package:flutter/material.dart';
-import 'model/user.dart';
-
-class LoginForm extends StatefulWidget {
-  @override
-  _LoginFormState createState() => _LoginFormState();
-}
-
-class _LoginFormState extends State<LoginForm> {
-  final _formKey = GlobalKey<FormState>();
-
-  String _email = '';
-  String _password = '';
-  String _name = '';
-  String _id= '';
-
-  void _submitForm() {
-    if (_formKey.currentState!.validate()) {
-      // If the form is valid, create a User object and log in
-      ApplicationUser user = ApplicationUser(id: _id, name: _name, email: _email, password: _password);
-      _login(user);
-    }
-  }
-
-  void _login(ApplicationUser user) {
-    // Code to log in the user goes here
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          TextFormField(
-            decoration: const InputDecoration(
-              labelText: 'Email',
-            ),
-            validator: (value) {
-              if (value == null || value.isEmpty) {
-                return 'Please enter your email address';
-              }
-              return null;
-            },
-            onSaved: (value) {
-              _email = value!;
-            },
-          ),
-          TextFormField(
-            obscureText: true,
-            decoration: const InputDecoration(
-              labelText: 'Password',
-            ),
-            validator: (value) {
-              if (value == null || value.isEmpty) {
-                return 'Please enter your password';
-              }
-              return null;
-            },
-            onSaved: (value) {
-              _password = value!;
-            },
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 16.0),
-            child: RaisedButton(
-              onPressed: _submitForm,
-              child: Text('Log in'),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}*/
-
